better reconnect handling: make sure to update message list
The new feed offers a way to request missing messages (=messages
with timestamps greater than the timestamp of the latest message
we already have). Sort into the message list.
There is a problem: the coloring scheme only works for even number of
new messages, but not really a big deal.

Refactor the code at the same time (too much of code duplication).

Also store the time stamp of messages in the list view in their "tags".