Overwrite properties only if "OK" is pressed
If the properties are overwritten immediately, the UI already reacts
to changes while the dialog is still open (and potentially is canceled).

Use ValidateChildren() on OK click to actually write to properties.
On Cancel the properties are reloaded - not nice, but works. Not
really needed, as the properties in fact are not changed yet (not
validated), but I did not find a way to resync the databinding.