use wtListener as AS interface and simplify code a bit
GetPlanes now just sends the UDP message (ASSETPATH) to AS and then
blocks on waitHandle. This waitHandle is set by wtListener once
it receives the right message.

Previously the "progress" of the BackgroundWorker had to handle the
UDP message, now only the dxcall is passed. The processing of the messages
is handled internally in AirScoutIF.

The BackgroundWorker should probably be removed.