compute list of displayed calls and pass to AS request
Getplanes is called for a list of calls that are currently shown.
This list is assembled in fill_AS_list() - called every time the
listview is refilled, the window is resized or scrolled.

The check if AS needs to be checked is moved into fill_AS_list(), if too
close or far, the call is not put to the list. Same for mycall.