If a connection fails, code was blocking for good
Fix error in tw lib, if a connection could not be established, the
code would get stuck on a wait event for good.

This appeared to make the autoconnect fail - but in fact it was just a
symptom: if autoconnect had to kick in because of a connection failure
the next connection may fail, too, triggering the errornous state.