refactor building AS_list and use for watchlist
AS_list is populated with the calls from the user list that are currently
visible. We do this by iterating over the items in lv_Calls list and
checking if they are currently visible or not.

All entries that have the right distance to be useful for Airplane scatter
are added to the watchlist, too.

Watchlist is sent when appropriate but not build on the fly but toegether
with AS_list now. That avoids useless entries that are not displayed
anymore in the user list.