fix "enumarate" problem in Check_QSOs()
Locking alone does not help, as the timer for the UI runs in the
UI thread - so taking the lock does not "lock" the code running in
the timer from the regular code. Use a "separate" bool. But as the code
runs in the same thread we need to skip the code, otherwise it would
deadlock.