rework scheme to handle local qrv DB
The local DB is meant as an overlay to the station DB.
So we need a local DB entry for every call in the user list.
Copy/merge from station DB and modify the local DB if needed.

Store only entries, that are not "QRV_STATE.unknown".
Worked is treated as QRV.

set TIME to "now" for calls that are new. Otherwise rely on the incoming
messages to update the TIME.