wtLogSync: Updates (update a QSO, prefer "owner", dispose())
We do not care about mode. Changing band is not supported.
Deleting a QSO also not.

Requesting QSOs from stations:
Prefer "owner", if there is a timeout, find an alternative.
Still not ideal... but seems to work and can cope with packet loss

Add debug option to simulate packet loss

Also: only change wtStationSyncList if anything actually changed. May
help to reduce the number of exceptions because enumaration changed in
a foreach

add Dispose() method to log code:
Allows to stop the UDP broadcast receiver in wtLogSync.