FIX: Win-Test network sync does not work sometimes
In fact it was dependend on what data was in the Win-Test datagrams,
e.g. if you used "OP", updating QSO data failed.

It is caused by using Split() with StringSplitOptions.RemoveEmptyEntries
option - which does not make sense and skips empty parts. Thus
screwing up the index of the strings.

Also make ADDQSO a bit more robust.