fix old messages received right after connect
After connect we receive a number of old messages that unfortunately are
not sorted by time.

Instead of putting them into the listview in the correct order with some
magic, put them all in a separate list in KSTcom and fed them into
the code that fills the list en bloc.

That simplifies the processing in the UI a lot, also for messages directed
to me (lower window).

Also fixes the issue that due to the new "SERVER" message sent old
messages were suppressed currently.