New AS visualization UI
In addition to the "normal" ToolTip add an AS specific tooltip.
It will stay a bit longer and will not be interrupted by the regular
ToolTip.

ShowASToolTip() interprets the string passed as a formatted list of planes
and renders them graphically (in addition to the text) the same way as
the AS column (so colored elipse sized to the category of the plane etc.)

This additional information is available in the message windows, too.
There the current status of planes for a call is visualed in the context
menu and you can either launch the detailed view in Airscout or display
the list of planes.