Use the Win-Test UDP port number also for sending
Win-Test uses its port number as source and destination port.
Do the same (in the hope to not get dropped by Wifi APs with limited
resources).

This requires some refactoring to the code as you cannot open the socket
twice with same port, so you need to share one socket.
This is opened and handled inside WinTest.cs