fix changes to the "bands" columns do not trigger user filter updates
Calls in the user area are not filtered out if "worked" (green).
One has to trigger an update manually...

Fix by implementing a timer that is retriggered by any change to the user
area. In case the bands column get triggered, force a delay of 3s.
Otherwise it would be too difficult to toggle between red/black/grey when
calls are to be filtered out.